Write derived run display JSON in helm-summarize instead of helm-run
